requirements.txt
aiohttp==3.13.1
blinker==1.9.0
boto3==1.35.29
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
Flask==3.1.2
flask-cors==5.0.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.3
packaging==25.0
pandas==2.3.3
Pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
quart==0.19.6
requests==2.32.5
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3






